requirements.txt
streamlit>=1.55.0
pandas>=2.3.3
numpy>=2.2.6
opencv-python>=4.13.0.92
Pillow>=12.1.1
pytesseract>=0.3.13
pdf2image>=1.17.0
pypdf>=3.17.4
pdfplumber>=0.11.9
camelot-py>=1.0.9
python-dotenv>=1.2.2
google-cloud-documentai>=3.11.0
chardet>=5,<6
